Rename Tagstore API app
