+++ blockstore/urls.py
@@ -34,7 +34,7 @@ urlpatterns = auth_urlpatterns + [
     url(r'^api-auth/', include(auth_urlpatterns, namespace='rest_framework')),
     url(r'^auto_auth/$', core_views.AutoAuth.as_view(), name='auto_auth'),
     url(r'^health/$', core_views.health, name='health'),
-    url(r'^tagstore/', include('tagstore.tagstore_rest.urls', namespace='tagstore')),
+    url(r'^tagstore/', include('tagstore.api.urls', namespace='tagstore')),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
 
 if settings.DEBUG:  # pragma: no cover
