+++ tagstore/api/apps.py
@@ -1,0 +1,13 @@
+"""
+Django app configuration for the Tagstore REST API
+"""
+from django.apps import AppConfig
+
+
+class TagstoreApiConfig(AppConfig):
+    """
+    Django app configuration for the Tagstore REST API
+    """
+    name = 'tagstore.api'
+    verbose_name = "Tagstore REST API"
+    label = 'tagstore_rest_api'
