+++ tagstore/api/urls.py
@@ -7,5 +7,5 @@ contain namespaces for the active versions of the API.
 from django.conf.urls import url, include
 
 urlpatterns = [
-    url(r'^api/v1/', include('tagstore.tagstore_rest.v1.urls', namespace='apiv1')),
+    url(r'^api/v1/', include('tagstore.api.v1.urls', namespace='apiv1')),
 ]
